refactor(filesystem): ♻  split file manager modules and extract watch refresh scheduler
